requirements/prod.txt
# Requirements to run Doozer
-e svn+http://code.djangoproject.com/svn/django/trunk@13302#egg=Django
-e git://github.com/clouserw/tower.git#egg=tower
-e git://github.com/jbalogh/schematic.git#egg=schematic
-e git://github.com/jbalogh/django-cache-machine.git#egg=django-cache-machine
python-memcached
django-registration==0.7
